Update submit_bash.sh to include HFS2d training command and comment out WNO2d command
This commit adds a new command to train the HFS2d model using the specified configuration file while commenting out the existing command for the WNO2d model. This change facilitates the training process for the new model while maintaining the option to run the WNO2d training in the future.